fix(federated-unet): stringify torch.__version__ in get_status, bump to 0.1.1
Returning torch.__version__ from a service method fails as "ModuleNotFoundError: No module named 'torch'" raised inside Ray's deserializer, because it is a torch-defined str subclass and the BioEngine proxy that unpickles the reply has no torch installed. It reads exactly like a broken runtime env and is not one.

Also casts the per-image Dice and IoU to float so the metrics record carries no numpy scalars, and adds deploy.py, the three-instance deploy helper.

+++ apps/federated-unet/training.py
@@ -110,10 +110,10 @@ def evaluate(
         per_image.append(
             {
                 "index": index,
-                "dice": (2.0 * intersection / (prediction.sum() + truth.sum()))
+                "dice": float(2.0 * intersection / (prediction.sum() + truth.sum()))
                 if (prediction.sum() + truth.sum()) > 0
                 else 1.0,
-                "iou": (intersection / union) if union > 0 else 1.0,
+                "iou": float(intersection / union) if union > 0 else 1.0,
                 "pred_foreground_fraction": float(prediction.mean()),
                 "true_foreground_fraction": float(truth.mean()),
             }
